+++ CLAUDE.md
@@ -1,1 +1,1 @@
-- All test unitest should be save in: invenio_notify_test
+- All test unitest should be save in: tests
