+++ tests/fixtures/endorsement_fixture.py
