+++ tests/test_endorsement_reply.py
@@ -2,10 +2,10 @@ import pytest
 
 from invenio_notify.records.models import EndorsementReplyModel, EndorsementRequestModel
 from invenio_notify.proxies import current_endorsement_reply_service
-from invenio_notify_test.fixtures.inbox_fixture import create_inbox
-from invenio_notify_test.fixtures.actor_fixture import create_actor
-from invenio_notify_test.fixtures.endorsement_request_fixture import create_endorsement_request
-from invenio_notify_test.fixtures.endorsement_reply_fixture import create_endorsement_reply
+from tests.fixtures.inbox_fixture import create_inbox
+from tests.fixtures.actor_fixture import create_actor
+from tests.fixtures.endorsement_request_fixture import create_endorsement_request
+from tests.fixtures.endorsement_reply_fixture import create_endorsement_reply
 
 
 def test_model_create(create_endorsement_reply):
